Refactor nzbToMediaUserScripts -> user_scripts
